add comment

+++ src/main/java/com/example/travel/controller/QuestionController.java
@@ -14,7 +14,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name="id") Integer id,
+    public String question(@PathVariable(name="id") Long id,
                            Model model) {
         questionService.incView(id);
         QuestionDTO questionDTO = questionService.getById(id);
